updated server and client connection
server reworked to use a list to store only move data,  which is initialized to None and reset to None once move data is sent. Avoids de-sync issue of connections being in separate threads.

Update data sent when clients are waiting on a move. Server checks whether the move list has data for the opposing player to send. Otherwise server and clients are waiting for user to use Pygame UI to make a move